+++ CHANGES.md
@@ -20,6 +20,8 @@
     - `GRI`: `DSWI4`
     - `WV_SI`: `NDSIWV`
     - `PANI`: `BITM`
+- **ENH: Adding the support of WorldView Legion constellation** ([#183](https://github.com/sertit/eoreader/issues/183))
+
 
 ## 0.21.10 (2025-02-18)